Add full_config to compute client args.
This change adds the full config to functions computing the client args.
This is added now to compute the configured endpoint URL,
but could be used for future features that may read from a different
config file section.

This change also adds in the configured endpoint resolution when computing the
args and all tests related to computing the configured endpoint URL.