Change verb of opt out flag.
This changes the opt out flag from a positive "use" with a default of
True to a negative "ignore" with a default of False.

The logic for computing if it should be ignored was simplified by
setting the config store property default to False. Now it only needs
to check if it was set first in the client config and not check for
another None value before returning it from the config store.